fix: move newsletter to own page
